(#15) feature(client/home): add home header
